Create README.md
This Script audits all of the Items within a Geodatabase. The first step of the process is to accesses the GDB_ITEMS table, and export the table to a scratch gdb. Then Definition and Documentation fields are deleted because excel does not like them. Then the process accesses the GDB_ITEMTYPES table, and export the table to a scratch gdb. Then Join Field Management Geoprocessing tool is run to bring the two tables together then exported to excel.